BUG: Fix two bugs.
The first bug was reported in issue gh-4.  `odeintw` didn't handle
the tfirst argument.

A second bug was found while adding support for tfirst.  In the
complex case, if the user's function returned a *scalar* instead
of a 1-d array of length 1, an exception would be raised because
numpy doesn't allow creating a float64 view of a complex scalar.

Closes gh-4.